gitlab_api/requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.3.2
DateTime==5.5
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
python-gitlab==4.12.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.0.3
